backend/pyproject.toml
[project]
name = "monet-backend"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "google-adk>=1.20.0",
    "google-genai>=1.44.0",
    "fastapi>=0.115.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.32.0",
    "asyncpg>=0.30.0",
    "google-cloud-storage>=2.18.0",
    "python-multipart>=0.0.9",
    "cloud-sql-python-connector[asyncpg]>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
